add support for specifying name of output notebook
If the 'new-notebook' option is given, then a new kernel is
started and the cells get appended to a new notebook. The name
may be optionally provided. This also affects the naming of the
cell output files.

The kernel may also be specified with the 'kernel' option, and this
is only legal if 'new-notebook' also occurs in the same directive.