Add direct dependency on ipykernel
ipywidgets 8.1.0 has been released, which includes the change at
https://github.com/jupyter-widgets/ipywidgets/pull/3811.
With this change, ipykernel is no longer a dependnecy of ipywidgets,
so jupyer-sphinx must now depend on ipykernel directly rather
than assume it will be installed as a transitive dependency of
ipywidgets.  Without this change, jupyter-sphinx tests will fail CI.
This has additionally led to CI failures in some packages that
depend on jupyter-sphinx but not ipykernel.